requirements.txt
Appium-Python-Client>=3.1.0
selenium>=4.0.0
flask
requests
python-dotenv
openai
fastapi
uvicorn
slowapi
httpx
anthropic>=0.18.1
apache-airflow-providers-redis
